final_product/requirements.txt
blinker==1.8.1
click==8.1.7
ecdsa==0.19.0
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
pillow==10.3.0
pyzbar==0.1.9
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
